leadmanager/requirements.txt
Package             Version
------------------- -------
asgiref             3.5.0  
Django              3.2.12 
djangorestframework 3.13.1 
pip                 10.0.1 
pytz                2021.3 
setuptools          39.0.1 
sqlparse            0.4.2  
typing-extensions   4.0.1  
